[FAB-8722] Pre-create Docker CouchDB system databases
Problem:

Startup of CouchDB docker images causes large numbers of errors to be
displayed if the peer is not started.  The peer automatically creates
system databases automatically.  If system databases are missing, errors
will be displayed frequently.

Solution:
Pre-create the CouchDB system databases as part of the CouchDB docker image.

Change-Id: Ia18474c6f1bd924eb90026196eab3e42047550d3
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>