[FAB-10421] Username improvements for CouchDB image
Currently, the admin username and password are appended to the
docker.ini file in etc/local.d on container start/restart.  This
can cause extra [admins] stanzas to appear in the docker.ini.
Also, a default local.ini file is created by CouchDB on initial
startup in the /etc directory.

Storing the user name and password in the docker.ini is confusing
since there is an [admins] stanza in the local.ini.

This change moves the management of the username and password in
the local.ini.

Also, a volume is being added for the for the /etc directory
to allow easier backup and maintenance of configuration files
across server upgrades.

Change-Id: Ie5280a81547b131b829c724e2342fd386d1adc54
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>